[Scala 3] Add support for given imports
Previously, organize-imports-rule would throw an exception in case of given imports, now it should sort them correctly.

Most of the existing rules should apply correctly to imports with givens, with an exception that grouped imports will be separated, so that in case of importing a type needed for given we will not get an error.